fix: run getContentNode if using front page to properly hydrate Apollo cache
This fixes usePost() not returning anything if using a frontpage and client-side queries are disabled. Also, extracted out query into nextFetchFromWP()
